Changed Kernel->Add
This commit removes NVMeFix.kext.
Initially, I wanted to boot macOS on NVME drive. However, the SATA
controller which controls 2 NVME drives (1 for Windows) and 2 HDD drives
is configured to RAID mode (for 2 HDD drives).

Therefore, macOS did not recognize the NVME drives. However, it
recognized Kingston SSD which was connected to external SATA controller
which was configured to AHCI mode.

There are 2 options now:
  1. Change the On-board SATA controller from VMD (RAID) to AHCI mode.
  2. Install macOS on Kingston drive.

This commit chooses the second option.
Please add NVMeFix.kext to EFI\OC\Kexts\ if you want to revert the commit.

Signed-off-by: Loc Pham <pploc.1212216@gmail.com>